Use correct buffer size in string subscript operator.

+++ src/wren_value.c
@@ -341,7 +341,7 @@ Value wrenNewUninitializedString(WrenVM* vm, size_t length)
 {
   ObjString* string = allocate(vm, sizeof(ObjString) + length + 1);
   initObj(vm, &string->obj, OBJ_STRING, vm->stringClass);
-  string->length = length;
+  string->length = (int)length;
 
   return OBJ_VAL(string);
 }
